Fix typescript errors

+++ src/modal/delete.tsx
@@ -1,32 +1,37 @@
-import { useLiveQuery } from 'dexie-react-hooks';
-import { useMemo } from 'react';
-import { db } from '../lib/database';
-import { MODAL_DISMISS_TIMEOUT_MS } from '../lib/modal';
-import useSettings from '../store/settings';
+import { useLiveQuery } from "dexie-react-hooks";
+import { useMemo } from "react";
+import { MODAL_DISMISS_TIMEOUT_MS } from "../lib/const";
+import db from "../lib/database";
+import { AssistantListing } from "../lib/types";
+import useSettings from "../store/settings";
+import { closeModal } from "../lib/modal";
 
-export const DeleteAssistantModalId = 'deleteAssistantModal';
+export const DeleteAssistantModalId = "deleteAssistantModal";
 
 export default function DeleteAssistantModal() {
-  const assistants = useLiveQuery(async () =>
+  const assistants: AssistantListing[] | undefined = useLiveQuery(async () =>
     (await db.assistant.toArray()).map((a, i) => ({ ...a, index: i + 1 }))
   );
 
   const settingsStore = useSettings();
-  const activeAssistant = useMemo(
+  const activeAssistant: AssistantListing | undefined = useMemo(
     () => assistants?.find((a) => a.id === settingsStore.activeAssistantId),
     [assistants, settingsStore.activeAssistantId]
   );
 
   const onDelete = async () => {
+    if (!activeAssistant || !assistants) return;
     try {
       settingsStore.setActiveAssistant(undefined);
-      const remainingAssistants = assistants.filter((a) => a.id != activeAssistant.id);
+      const remainingAssistants = assistants?.filter(
+        (a) => a.id != activeAssistant?.id
+      );
       await db.assistant.where({ id: activeAssistant.id }).delete();
       if (remainingAssistants.length > 0) {
         settingsStore.setActiveAssistant(remainingAssistants[0].id);
       }
       setTimeout(() => {
-        document.getElementById(DeleteAssistantModalId).close();
+        closeModal(DeleteAssistantModalId);
       }, MODAL_DISMISS_TIMEOUT_MS);
     } catch (_error) {
       return;
@@ -37,14 +42,17 @@ export default function DeleteAssistantModal() {
     <dialog id={DeleteAssistantModalId} className="modal">
       <div className="modal-box">
         <div className="text-center mb-6">
-          <h3 className="font-bold text-lg">Delete '{activeAssistant?.name}'?</h3>
+          <h3 className="font-bold text-lg">
+            Delete '{activeAssistant?.name}'?
+          </h3>
         </div>
         <div>
           <div>
             <p>
-              Just a heads up - once you say goodbye, they'll be gone for good and{' '}
-              <kbd className="kbd">ALT</kbd>
-              <kbd className="kbd">{activeAssistant?.index}</kbd> shortcut will be reassigned!
+              Just a heads up - once you say goodbye, they'll be gone for good
+              and <kbd className="kbd">ALT</kbd>
+              <kbd className="kbd">{activeAssistant?.index}</kbd> shortcut will
+              be reassigned!
             </p>
           </div>
           <div className="modal-action flex">
@@ -54,7 +62,11 @@ export default function DeleteAssistantModal() {
               </button>
             </form>
             <div className="flex-1">
-              <button type="button" className="w-full btn btn-error" onClick={onDelete}>
+              <button
+                type="button"
+                className="w-full btn btn-error"
+                onClick={onDelete}
+              >
                 Delete
               </button>
             </div>
